+++ packages/react/stories/imdb/stories/greyVest.js
@@ -121,6 +121,8 @@ let tree = Contexture({
       key: 'results',
       type: 'results',
       include: [
+        'imdbId',
+        'runtimeMinutes',
         'poster',
         'title',
         'actors',
@@ -339,7 +341,10 @@ export default () => (
                 <div className="gv-box">
                   <ResultTable
                     path={['root', 'results']}
-                    fields={schemas[tree.tree.schema].fields}
+                    fields={_.omit(
+                      ['imdbId', 'runtimeMinutes'],
+                      schemas[tree.tree.schema].fields
+                    )}
                     criteria={['root', 'criteria']}
                     typeComponents={TypeMap}
                   />